[tests] Refactor tests
Move unit tests inside packages that match the package of the unit under test.
