+++ com.reprezen.swagedit.tests/src/com/reprezen/swagedit/json/SwaggerSchemaTest.xtend
@@ -8,7 +8,7 @@
  *  Contributors:
  *     ModelSolv, Inc. - initial API and implementation and/or initial documentation
  *******************************************************************************/
-package com.reprezen.swagedit.tests
+package com.reprezen.swagedit.json
 
 import com.reprezen.swagedit.json.JsonSchemaManager
 import org.junit.Test
